language: make exact scalar spellings the only numeric surface
Use one stable scalar registry for lexer, analyzer, runtime, IR, AOT, formatter, LSP, and generated knowledge consumers. Remove the retired int, byte, and float names together with global conversion calls, name-based type fallbacks, and the strconv module.

Keep numeric conversions explicit through as, expose whole-input text parsing through i64 and f64 type namespaces, migrate the source corpus, and reject every removed compiler-owned identity without reserving the retired identifiers for user code.

+++ tests/aot/run_native_debug_smoke.py
@@ -64,7 +64,7 @@ DWARFDUMP_CANDIDATES = (
 # report "no source locals" for what is really a fixture bug).
 BREAK_MARKER = "// BREAK"
 
-SCALAR_SOURCE = """fn compute(seed: int) -> int {
+SCALAR_SOURCE = """fn compute(seed: i64) -> i64 {
     if (seed <= 0) { return 0 }
     var answer = seed + 1
     var doubled = answer * 2
@@ -95,10 +95,10 @@ make(20)
 
 # No bare `yield`: that statement form was removed with the go/defer surface
 # convergence, and the fixture had kept it, so this scenario failed to parse.
-CORO_SOURCE = """fn produce(seed: int) -> int {
+CORO_SOURCE = """fn produce(seed: i64) -> i64 {
     return seed + 1
 }
-fn worker(seed: int) -> int {
+fn worker(seed: i64) -> i64 {
     var task = go produce(seed)
     var answer = await task
     var doubled = answer * 2
